+++ vimiv/gtk.py
@@ -20,6 +20,7 @@ from vimiv import imageactions
 from vimiv.completions import Completion
 from vimiv.toggle import FullscreenToggler, VariableToggler
 from vimiv.commands import Commands
+from vimiv.commandline import CommandLine
 
 # Directories
 vimivdir = os.path.join(os.path.expanduser("~"), ".vimiv")
@@ -86,6 +87,7 @@ class Vimiv(Gtk.Window):
         self.toggle_fullscreen = FullscreenToggler(self, settings)
         self.toggle_vars = VariableToggler(self, settings)
         Commands(self)
+        self.commandline = CommandLine(self)
 
     def delete(self):
         """ Delete all marked images or the current one """
@@ -1399,100 +1401,6 @@ class Vimiv(Gtk.Window):
                 self.grid.set_size_request(self.winsize[0], 10)
             self.cmd_line_info.set_max_width_chars(self.winsize[0]/16)
 
-    def cmd_handler(self, entry):
-        """ Handles input from the entry, namely if it is a path to be focused
-        or a (external) command to be run """
-        # cmd from input
-        command = entry.get_text()
-        # And close the cmd line
-        self.cmd_line_leave()
-        if command[0] == "/":  # Search
-            self.search(command.lstrip("/"))
-        else:  # Run a command
-            cmd = command.lstrip(":")
-            # If there was no command just leave
-            if not cmd:
-                return
-            # Parse different starts
-            if cmd[0] == "!":
-                self.run_external_command(cmd)
-            elif cmd[0] == "~" or cmd[0] == "." or cmd[0] == "/":
-                self.cmd_path(cmd)
-            else:
-                self.num_str = ""  # Be able to repeat commands
-                while True:
-                    try:
-                        num = int(cmd[0])
-                        self.num_str = self.num_str + str(num)
-                        cmd = cmd[1:]
-                    except:
-                        break
-                self.run_command(cmd)
-        # Save the cmd to a list
-        if command in self.cmd_history:
-            self.cmd_history.remove(command)
-        self.cmd_history.insert(0, command)
-        self.cmd_pos = 0
-
-    def run_external_command(self, cmd):
-        """ Run the entered command in the terminal """
-        # Check on which file(s) % and * should operate
-        if self.last_focused == "lib" and self.files:
-            filelist = self.files
-            fil = self.files[self.treepos]
-        elif self.paths:
-            filelist = self.paths
-            fil = self.paths[self.index]
-        else:
-            filelist = []
-            fil = ""
-        if self.marked:  # Always operate on marked files if they exist
-            filelist = self.marked
-        # Escape spaces for the shell
-        fil = fil.replace(" ", "\\\\\\\\ ")
-        for i, f in enumerate(filelist):
-            filelist[i] = f.replace(" ", "\\\\\\\\ ")
-        cmd = cmd[1:]
-        # Substitute % and * with escaping
-        cmd = re.sub(r'(?<!\\)(%)', fil, cmd)
-        cmd = re.sub(r'(?<!\\)(\*)', " ".join(filelist), cmd)
-        cmd = re.sub(r'(\\)(?!\\)', '', cmd)
-        # Run the command in an extra thread
-        dir = os.path.abspath(".")
-        cmd_thread = Thread(target=self.thread_for_external, args=(cmd, dir))
-        cmd_thread.start()
-        # Undo the escaping
-        fil = fil.replace("\\\\\\\\ ", " ")
-        for i, f in enumerate(filelist):
-            filelist[i] = f.replace("\\\\\\\\ ", " ")
-
-    def thread_for_external(self, cmd, dir):
-        """ Starting a new thread for external commands """
-        try:
-            # Possibility to "pipe to vimiv"
-            if cmd[-1] == "|":
-                cmd = cmd.rstrip("|")
-                p = Popen(cmd, stdout=PIPE, stderr=PIPE, shell=True)
-                pipe = True
-            else:
-                p = Popen(cmd, stderr=PIPE, shell=True)
-                pipe = False
-            # Get output and error and run the command
-            out, err = p.communicate()
-            if p.returncode:
-                err = err.decode('utf-8').split("\n")[0]
-                self.err_message(err)
-            else:
-                GLib.timeout_add(1, self.reload_changes, dir, True, pipe, out)
-            # Reload everything after an external command if we haven't moved,
-            # you never know what happend ...
-            # Must be in a timer because not all Gtk stuff can be accessed from
-            # an external thread
-        except FileNotFoundError as e:
-            e = str(e)
-            cmd = e.split()[-1]
-            self.err_message("Command %s not found" % (cmd))
-
     def reload_changes(self, dir, reload_path=True, pipe=False, input=None):
         """ Reload everything, meaning filelist in library and image """
         if (dir == os.path.abspath(".") and dir != tagdir and
@@ -1516,119 +1424,8 @@ class Vimiv(Gtk.Window):
                     self.thumb_reload(image, i, False)
         # Run the pipe
         if pipe:
-            self.pipe(input)
+            self.commandline.pipe(input)
         return False  # To stop the timer
-
-    def pipe(self, input):
-        """ Run output of external command in a pipe
-            This checks for directories, files and vimiv commands """
-        # Leave if no input came
-        if not input:
-            self.err_message("No input from pipe")
-            return
-        # Make the input a file
-        input = input.decode('utf-8')
-        input = input.split("\n")[:-1]  # List of commands without empty line
-        startout = input[0]
-        # Do different stuff depending on the first line of input
-        if os.path.isdir(startout):
-            self.move_up(startout)
-        elif os.path.isfile(startout):
-            # Remember oldfile if no image was in filelist
-            if self.paths:
-                old_pos = [self.paths[self.index]]
-                self.paths = []
-            else:
-                old_pos = []
-            # Populate filelist
-            self.paths, self.index = populate(input)
-            if self.paths:  # Images were found
-                self.scrolled_win.show()
-                self.move_index(False, False, 0)
-                # Close library if necessary
-                if self.library_toggled:
-                    self.toggle_library()
-            elif old_pos:  # Nothing found, go back
-                self.paths, self.index = populate(old_pos)
-                self.err_message("No image found")
-        else:
-            # Run every line as an internal command
-            for cmd in input:
-                self.run_command(cmd)
-
-    def cmd_path(self, path):
-        """ Run a path command, namely populate files or focus directory """
-        # Expand home
-        if path[0] == "~":
-            path = os.path.expanduser("~") + path[1:]
-        try:
-            path = os.path.abspath(path)
-            if not os.path.exists(path):
-                raise ValueError
-            elif os.path.isdir(path):
-                self.move_up(path)
-                # If we open the library it must be focused as well
-                self.last_focused = "lib"
-            else:
-                # If it is an image open it
-                self.paths = []
-                self.paths, index = populate([path])
-                self.index = 0
-                self.move_index(True, False, index)
-                #  Reload library in lib mode, do not open it in image mode
-                abspath = os.path.dirname(path)
-                if self.last_focused == "lib":
-                    self.last_selected = path
-                    self.move_up(abspath)
-                    # Focus it in the treeview so it can be accessed via "l"
-                    for i, fil in enumerate(self.files):
-                        if fil in path:
-                            self.treeview.set_cursor(Gtk.TreePath(i),
-                                                     None, False)
-                            self.treepos = i
-                            break
-                    # Show the image
-                    self.grid.set_size_request(self.library_width-
-                                               self.border_width, 10)
-                    self.scrolled_win.show()
-                else:
-                    self.move_up(abspath, True)
-        except:
-            self.err_message("Warning: Not a valid path")
-
-    def run_command(self, cmd):
-        """ Run the correct internal cmd """
-        parts = cmd.split()
-        if "set" in cmd:
-            arg = parts[2:]
-            cmd = " ".join(parts[:2])
-        else:
-            arg = parts[1:]
-            cmd = parts[0]
-        # Check if the command exists
-        if cmd in self.commands.keys():  # Run it
-            function = self.commands[cmd][0]
-            default_args = self.commands[cmd][1:]
-            arg = default_args + arg
-            # Check for wrong arguments
-            try:
-                if arg:
-                    function(*arg)
-                else:
-                    function()
-            except TypeError as e:
-                self.err_message(str(e))
-            except SyntaxError:
-                err = ("SyntaxError: are all strings closed " +
-                       "and special chars quoted?")
-                self.err_message(err)
-            except NameError as e:
-                argstr = "('" + arg + "')"
-                arg = "(" + arg + ")"
-                function = function.replace(arg, argstr)
-                exec(function)
-        else:  # Through an error
-            self.err_message("No such command: %s" % (cmd))
 
     def history(self, down):
         """ Update the cmd_handler text with history """
@@ -1961,7 +1758,7 @@ class Vimiv(Gtk.Window):
         # Command line
         self.cmd_line_box = Gtk.HBox(False, 0)
         self.cmd_line = Gtk.Entry()
-        self.cmd_line.connect("activate", self.cmd_handler)
+        self.cmd_line.connect("activate", self.commandline.cmd_handler)
         self.cmd_line.connect("key_press_event",
                               self.handle_key_press, "COMMAND")
         self.cmd_line.connect("changed", self.cmd_check_close)
